(update) rename CLICommand --> Command
